+++ src/main/java/net/sargue/mailgun/Mail.java
@@ -1,15 +1,7 @@
 package net.sargue.mailgun;
 
-import org.glassfish.jersey.client.JerseyClientBuilder;
-
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.client.Invocation;
-import javax.ws.rs.client.InvocationCallback;
+import javax.ws.rs.client.*;
 import java.util.List;
-
-import static org.glassfish.jersey.client.ClientProperties.CONNECT_TIMEOUT;
-import static org.glassfish.jersey.client.ClientProperties.READ_TIMEOUT;
 
 /**
  * Representation of a Mailgun's mail request.
@@ -136,24 +128,13 @@ public abstract class Mail {
 
     abstract void prepareSend();
 
-    void configureClient(Client client) {
+    void configureTarget(WebTarget target) {
         //defaults to no-op
     }
 
     private Invocation.Builder request() {
-        Client client = JerseyClientBuilder.newClient();
-
-        if (configuration.connectTimeout() != 0)
-            client.property(CONNECT_TIMEOUT, configuration.connectTimeout());
-        if (configuration.readTimeout() != 0)
-            client.property(READ_TIMEOUT, configuration.readTimeout());
-
-        configureClient(client);
-        return client
-                .register(configuration.httpAuthenticationFeature())
-                .target(configuration.apiUrl())
-                .path(configuration.domain())
-                .path("messages")
-                .request();
+        WebTarget target = configuration.getTarget();
+        configureTarget(target);
+        return target.path(configuration.domain()).path("messages").request();
     }
 }